src/url_shortener/shortener/serializers.py
from rest_framework import serializers

from shortener.models import ShortenedURL


class ShortenedURLSerializer(serializers.ModelSerializer):
    class Meta:
        model = ShortenedURL
        fields = ['id', 'original_url', 'short_code']
